Professional Portfolio
In this challenge, we needed to make a professional portfolio website from scratch. 
The links must be functional with direction to the appropriate links. 
The repository has a unique name. I went with a short and direct name ProPort.
To the best of my knowledge, the Repository follows best practices for file structure and naming conventions.
The repository follows best practices for class/id naming conventions, indentation, quality comments, etc.
The repository contains multiple descriptive commit messages.
This repository contains quality readme with description, screenshot, link to a deployed application.
The application resembles the mock-up functionality provided in the Challenge instructions.